pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyELIJAH"
version = "0.0.1"
authors = [
    { name = "Francesco Amadori", email = "francesco.a97.ing@outlook.it" },
    { name = "Paula Manuela Leguizamon Pineda", email = "paulamanuela.leguizamonpineda@studenti.unibo.it" },
    { name = "Leonardo Dal Cin", email = "leonardo.dalcin@studenti.unibo.it" },
    { name = "Tiziano Zingales", email = "tiziano.zingales@unipd.it" }
]
description = "A basic tool to detect (transit) and characterize exoplanets atmospheres"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "astropy",
    "contourpy",
    "PyYAML",
    "batman-package",
    "matplotlib",
    "pyqt5",
    "pandas",
    "sphinx",
    "pillow",
    "numpy==1.24.3",
]


classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
pyELIJAH = "pyELIJAH.__main__:main"

[project.urls]
Homepage = "https://github.com/francescoa97outlook/pyELIJAH"
Issues = "https://github.com/francescoa97outlook/pyELIJAH/issues"
